requirements.txt
Flask==2.3.2
gunicorn==21.2.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.101.0
gspread==6.1.2
